Optimised Multiple features following feedback

+++ backend/db/migrate/20210714140950_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[6.0]
   def change
     create_table :products do |t|
       t.string :name, null: false, unique: true, default: ""
-      t.float :price, null: false, default: 0
+      t.integer :price, null: false, default: 0
 
       t.timestamps
     end
