+++ backend/db/migrate/20210714141914_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration[6.0]
   def change
     create_table :orders do |t|
       t.references :user, null: false, foreign_key: true
-      t.float :total
+      t.integer :total
 
       t.timestamps
     end
